+++ nextstep-mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -57,7 +57,6 @@ public class DispatcherServlet extends HttpServlet {
             view.render(modelAndView.getModelMap(), requestContext);
         } catch (Exception e) {
             logger.error("Exception : {}", e);
-            e.printStackTrace();
         }
     }
