+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/argumentresolver/HandlerMethodArgumentResolver.java
@@ -5,5 +5,6 @@ import nextstep.mvc.tobe.RequestContext;
 
 public interface HandlerMethodArgumentResolver {
     boolean supports(MethodParameter methodParameter);
+
     Object resolve(RequestContext requestContext, MethodParameter methodParameter);
 }
