+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/view/View.java
@@ -2,8 +2,6 @@ package nextstep.mvc.tobe.view;
 
 import nextstep.mvc.tobe.RequestContext;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.util.Map;
 
 public interface View {
