+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/viewresolver/ViewResolver.java
@@ -4,5 +4,6 @@ import nextstep.mvc.tobe.view.View;
 
 public interface ViewResolver {
     boolean supports(String viewName);
+
     View resolve(String viewName);
 }
